requirements.txt
flask==3.0.3
python-telegram-bot==21.4
python-dotenv==1.0.1
requests==2.31.0
feedparser==6.0.11
aiohttp==3.9.5
beautifulsoup4==4.12.3
lxml==5.2.2